requirements.txt
faiss-cpu
sentence-transformers
numpy
rich
bs4
lxml
requests